clean up type warning in modern clang/gcc 32/64bit environment

+++ src/common/modes.h
@@ -3,7 +3,7 @@
 
 int is_channel (server *serv, char *chan);
 char get_nick_prefix (server *serv, unsigned int access);
-unsigned int nick_access (server *serv, char *nick, int *modechars);
+unsigned int nick_access (server *serv, char *nick, long *modechars);
 int mode_access (server *serv, char mode, char *prefix);
 void inbound_005 (server *serv, char *word[]);
 void handle_mode (server *serv, char *word[], char *word_eol[], char *nick, int numeric_324);
